Separated EmpaticaReader and VrCamp data
